Reverted milestoneRepository to versionRepository
