--- src/Repository/MilestoneRepository.php
@@ -1,41 +1,0 @@
-<?php
-
-namespace App\Repository;
-
-use App\Entity\Milestone;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Persistence\ManagerRegistry;
-
-/**
- * @extends ServiceEntityRepository<Milestone>
- *
- * @method Milestone|null find($id, $lockMode = null, $lockVersion = null)
- * @method Milestone|null findOneBy(array $criteria, array $orderBy = null)
- * @method Milestone[]    findAll()
- * @method Milestone[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
- */
-class MilestoneRepository extends ServiceEntityRepository
-{
-    public function __construct(ManagerRegistry $registry)
-    {
-        parent::__construct($registry, Milestone::class);
-    }
-
-    public function save(Milestone $entity, bool $flush = false): void
-    {
-        $this->getEntityManager()->persist($entity);
-
-        if ($flush) {
-            $this->getEntityManager()->flush();
-        }
-    }
-
-    public function remove(Milestone $entity, bool $flush = false): void
-    {
-        $this->getEntityManager()->remove($entity);
-
-        if ($flush) {
-            $this->getEntityManager()->flush();
-        }
-    }
-}
