+++ src/Repository/VersionRepository.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace App\Repository;
+
+use App\Entity\Version;
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Persistence\ManagerRegistry;
+
+/**
+ * @extends ServiceEntityRepository<Version>
+ *
+ * @method Version|null find($id, $lockMode = null, $lockVersion = null)
+ * @method Version|null findOneBy(array $criteria, array $orderBy = null)
+ * @method Version[]    findAll()
+ * @method Version[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ */
+class VersionRepository extends ServiceEntityRepository
+{
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, Version::class);
+    }
+
+    public function save(Version $entity, bool $flush = false): void
+    {
+        $this->getEntityManager()->persist($entity);
+
+        if ($flush) {
+            $this->getEntityManager()->flush();
+        }
+    }
+
+    public function remove(Version $entity, bool $flush = false): void
+    {
+        $this->getEntityManager()->remove($entity);
+
+        if ($flush) {
+            $this->getEntityManager()->flush();
+        }
+    }
+}
